added jumping / improved commenting
I have added a new option to the criteria file. The “jump” option
allows you to specify a location to jump to if the criteria is clicked.
Since ace.js doesn’t seem to allow multiline regex finding, I’ve used a
list of regex strings in the criteria for now. The last item in the
list is the “match” to highlight and any elements before that change
the start location to search. i.e. find the first one, then from there
find the second, then the third etc.