added meeting comments ui
Added a helper switch to note whether you have interviewed someone or
not. There is also a text area for you to type meeting comments in. You
could just type it directly into the script before, but this makes it
more consistent, which in turn helps with regex. The more Shark knows
about the script, the more it can do for you.